Introducing! Jacobs PseudoCode
Here is what I've got for the actual process. It needs a lot of work, but I think the bones are all there. Most of the stuff is commented out since I was working on it in my notebook and didn't want to screw too much up. Let's go over it soon, so we can get the tricky if statements figured out, and how we'll get inputs from the user to the program.